+++ light-curve/src/lib.rs
@@ -59,6 +59,7 @@ fn light_curve(py: Python, m: Bound<PyModule>) -> PyResult<()> {
     m.add_class::<f::BeyondNStd>()?;
     m.add_class::<f::Bins>()?;
     m.add_class::<f::BiweightScale>()?;
+    m.add_class::<f::Bootstrap>()?;
     m.add_class::<f::Chi2Pvar>()?;
     m.add_class::<f::ColorOfMaximum>()?;
     m.add_class::<f::ColorOfMedian>()?;